修改检视意见; Signed-off-by:liuhaosu@huawei.com

+++ services/scan_service/src/scan_service_utils.cpp
@@ -115,7 +115,7 @@ std::vector<std::string> ScanServiceUtils::ExtractIpOrPortFromUrl(const std::str
         end = url.find(delimiter, start);
     }
     tokens.push_back(url.substr(start));
-    if (tokens.size() < minTokenLength) {
+    if (tokens.size() < static_cast<size_t>(minTokenLength)) {
         SCAN_HILOGE("Url size < %{public}d ", minTokenLength);
         tokens.clear();
     }
